refactor: move useProfileFlags hook to state
